requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.1.0
astor==0.8.1
async-timeout==4.0.0
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.4b2
bleach==4.1.0
blessings==1.7
boto3==1.20.36
botocore==1.23.36
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
chumpy==0.70
click==8.0.3
cloudpickle==2.0.0
cmake==3.24.1.1
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.5.0
configparser==5.0.2
contextlib2==21.6.0
crypto==1.4.1
cryptography==35.0.0
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.5.0
decorator==5.1.0
deflicker==0.1.1
defusedxml==0.7.1
detectron2==0.6+cu113
discord==2.2.3
discord.py==2.2.3
dlib==19.24.0
docker-pycreds==0.4.0
docopt==0.6.2
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.3
et-xmlfile==1.1.0
face-alignment==1.3.5
ffmpeg-python==0.2.0
filelock==3.0.12
fire==0.4.0
flatbuffers==23.1.21
frozenlist==1.2.0
fsspec==2022.10.0
future==0.18.2
fvcore==0.1.5.post20211023
gdown==4.2.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==2.3.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.41.0
gspread==5.6.2
h5py==3.7.0
httplib2==0.21.0
huggingface-hub==0.10.1
humanfriendly==10.0
hyperopt==0.1.2
-e git+https://github.com/hhj1897/face_alignment.git@9cf5494e443f26d567972f3f50f6212d65b76c01#egg=ibug_face_alignment
-e git+https://github.com/hhj1897/face_detection.git@71852f00b815f568f3b51f045a418ae84cbe162a#egg=ibug_face_detection
-e git+https://github.com/hhj1897/face_pose_augmentation.git@79ad0457888e3b5f105f41fec00136e20a32feec#egg=ibug_face_pose_augmentation
idna==3.2
igraph==0.10.2
ImageAug==0.1.0.post0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
imutils==0.5.4
inplace-abn==1.1.1.dev3+ga8a3f82
iopath==0.1.9
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
json-tricks==3.15.5
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.2.1
kiwisolver==1.3.2
kornia==0.6.9
kubernetes==25.3.0
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
llvmlite==0.39.1
lmdb==1.2.1
lxml==4.6.4
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
matrix-client==0.4.0
mdurl==0.1.2
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.2
multidict==5.2.0
mypy-extensions==0.4.3
Naked==0.1.31
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nni==2.4
notebook==6.4.4
numba==0.56.4
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.1
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.13.1
onnxsim==0.4.33
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.68
openpyxl==3.0.9
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
portalocker==2.3.2
premailer==3.10.0
prettytable==2.1.0
progressbar2==4.2.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.0
psutil==5.8.0
ptflops==0.6.7
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.14.0
pyDeprecate==0.3.2
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.3.0
pymongo==3.12.0
pyparsing==2.4.7
pypng==0.20220715.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.7
python-utils==3.4.5
PythonWebHDFS==0.2.3
pytorch-lightning==1.6.0
pytorch-loss==0.0.0
pytorch-ranger==0.1.1
pytorch-toolbelt==0.5.0
pytorchcv==0.0.65
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.13.4
rich==13.5.0
rsa==4.7.2
s3transfer==0.5.0
schema==0.7.4
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
SecretStorage==3.3.1
Send2Trash==1.8.0
sentry-sdk==1.4.3
Shapely==1.7.1
shellescape==3.8.1
shortuuid==1.0.1
shyaml==0.6.2
simplejson==3.17.5
six==1.16.0
slack-cleaner2==3.1.0
slack-sdk==3.21.3
smmap==4.0.0
smplx==0.1.26
soupsieve==2.2.1
subprocess32==3.5.4
swin-window-process==0.0.0
sympy==1.11.1
tabulate==0.8.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
tflite-runtime==2.13.0
threadpoolctl==2.2.0
tifffile==2021.8.30
timm==0.4.5
toml==0.10.2
torch==1.13.0
torch-optimizer==0.1.0
torchaudio==0.13.0
torchgeometry==0.1.2
torchinfo==1.8.0
torchmetrics==0.10.2
torchvision==0.14.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
twilio==7.3.0
typing-extensions==4.4.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
wandb==0.12.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==9.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.1
yacs==0.1.8
yagmail==0.14.260
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
